requirements.txt
argparse
mysql-connector-python
pandas